Removed Utils File
Moved the contents of the utils file and removed the file since it
doesn't really make sense to have a utilities file with only one
utility.
