+++ dependency_file/Makefile
@@ -1,3 +1,3 @@
 $(PLUGIN_TARGET_DIR)/panda_$(PLUGIN_NAME).so: \
-	$(PLUGIN_OBJ_DIR)/$(PLUGIN_NAME).o \
-	$(PLUGIN_OBJ_DIR)/utils.o
+	$(PLUGIN_OBJ_DIR)/$(PLUGIN_NAME).o
+
