+++ dependency_file/dependency_file_def.h
@@ -282,6 +282,27 @@ void cbf_writeReturn(CPUState *cpu, target_ulong pc, uint32_t fd,
 std::string getFileName(CPUState *cpu, int fd, bool debug = false);
 
 /// <summary>
+/// Returns a string fetched from the specified memory address.
+/// </summary>
+/// <param name="cpu">
+/// The CPU State pointer.
+/// </param>
+/// <param name="maxSize">
+/// The maximum size of the string to be read from memory. If the actual string
+/// size exceeds this parameter, only the first maxSize number of characters
+/// will be returned. The actual string size may be smaller than this value, as
+/// this function will automatically stop when a null terminator character is
+/// encountered in memory.
+/// </param>
+/// <param name="addr">
+/// The address from which the string is to be read.
+/// </param>
+/// <returns>
+/// The guest's string at the memory location.
+/// </returns>
+std::string getGuestString(CPUState *cpu, size_t maxSize, target_ulong addr);
+
+/// <summary>
 /// Taints the contents of the buffer at the specified virtual address and of 
 /// the specified length. This function does nothing if taint2 is not currently
 /// enabled.
